cursor moving position with arrow keys

+++ kernel/include/kernel/hal/tty.h
@@ -19,8 +19,8 @@ struct tty_data {
     struct file *keyboard;
     struct key_event key_buffer;
     char *input_buffer;
-    int input_buffer_length;
-    int input_buffer_offset;
+    size_t input_buffer_length;
+    size_t input_buffer_offset;
     spinlock_t lock;
 };
